Added CONTRIBUTING.md, updated README.md
