+++ dependencies.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+sudo apt update
+sudo apt install build-essential -y 
+sudo apt install python3 -y
+sudo apt install libpoppler-cpp-dev pkg-config python-dev python3-pip -y
+sudo apt install figlet -y
+sudo apt install unoconv -y
+pip install pdftotext 
+sudo apt update
